refactor(semantic): extract `oxc_jsdoc` crate from `oxc_semantic` (#19706)
## Summary

- Extract JSDoc parsing code from `oxc_semantic` into a new `oxc_jsdoc` crate
- Move parser (`JSDoc`, `JSDocTag`), builder (`JSDocBuilder`), and all supporting types to the new crate
- Keep `JSDocFinder` in `oxc_semantic` since it depends on semantic types (`AstNode`, `AstNodes`)
- All existing imports via `oxc_semantic` continue to work through re-exports — no changes needed in `oxc_linter`
