+++ Cargo.toml
@@ -119,6 +119,7 @@ oxc_ecmascript = { version = "0.115.0", path = "crates/oxc_ecmascript" } # ECMAS
 oxc_estree = { version = "0.115.0", path = "crates/oxc_estree" } # ESTree format
 oxc_estree_tokens = { version = "0.115.0", path = "crates/oxc_estree_tokens" } # ESTree token conversion
 oxc_isolated_declarations = { version = "0.115.0", path = "crates/oxc_isolated_declarations" } # TS declaration generation
+oxc_jsdoc = { version = "0.115.0", path = "crates/oxc_jsdoc" } # JSDoc parsing
 oxc_mangler = { version = "0.115.0", path = "crates/oxc_mangler" } # Name mangling
 oxc_minifier = { version = "0.115.0", path = "crates/oxc_minifier" } # Code minification
 oxc_minify_napi = { version = "0.115.0", path = "napi/minify" } # Node.js minifier binding
