+++ crates/oxc_jsdoc/Cargo.toml
@@ -1,0 +1,27 @@
+[package]
+name = "oxc_jsdoc"
+version = "0.115.0"
+authors.workspace = true
+categories.workspace = true
+edition.workspace = true
+homepage.workspace = true
+include = ["/src"]
+keywords.workspace = true
+license.workspace = true
+publish = true
+repository.workspace = true
+rust-version.workspace = true
+description.workspace = true
+
+[lints]
+workspace = true
+
+[lib]
+doctest = false
+
+[dependencies]
+oxc_ast = { workspace = true }
+oxc_span = { workspace = true }
+
+rustc-hash = { workspace = true }
+
