+++ crates/oxc_jsdoc/src/lib.rs
@@ -1,0 +1,5 @@
+mod builder;
+pub mod parser;
+
+pub use builder::{JSDocBuilder, JSDocBuilderResult};
+pub use parser::{JSDoc, JSDocTag};
