+++ crates/oxc_jsdoc/src/parser/mod.rs
@@ -1,5 +1,5 @@
 mod jsdoc;
-mod jsdoc_parts;
+pub mod jsdoc_parts;
 mod jsdoc_tag;
 mod parse;
 mod utils;
