--- crates/oxc_semantic/src/jsdoc/mod.rs
@@ -1,7 +1,0 @@
-mod builder;
-mod finder;
-mod parser;
-
-pub use builder::JSDocBuilder;
-pub use finder::JSDocFinder;
-pub use parser::{JSDoc, JSDocTag};
